web/requirements.txt
aiokafka==0.11.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
captcha==0.6.0
certifi==2024.8.30
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
fastapi==0.112.2
future==1.0.0
h11==0.14.0
h2==4.1.0
hiredis==3.0.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.8
msgpack==1.0.8
nebula3-python==3.8.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-multipart==0.0.10
pytz==2024.1
redis==5.0.8
setuptools==74.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
uvicorn==0.30.6
websockets==13.0.1
